\docType{data}
\name{net1345}
\alias{net1345}
\title{Sample network topology using 1345 genes with interactions based on observed
protein-protein interaction networks.}
\description{
Sample network topology using 1345 genes with interactions
based on observed protein-protein interaction networks.
}
\author{
Guanghua Xiao \email{Guanghua.Xiao@UTSouthwestern.edu}

Jeffrey D. Allen \email{Jeffrey.Allen@UTSouthwestern.edu}
}

