\docType{data}
\name{net17}
\alias{net17}
\title{Sample network topology using 17 genes with interactions based on observed
protein-protein interaction networks.}
\description{
Sample network topology using 17 genes with interactions
based on observed protein-protein interaction networks.
}
\author{
Guanghua Xiao \email{Guanghua.Xiao@UTSouthwestern.edu}

Jeffrey D. Allen \email{Jeffrey.Allen@UTSouthwestern.edu}
}

