\docType{data}
\name{net44}
\alias{net44}
\title{Sample network topology using 44 genes with interactions based on observed
protein-protein interaction networks.}
\description{
  Sample network topology using 44 genes with interactions
  based on observed protein-protein interaction networks.
}
\author{
  Guanghua Xiao \email{Guanghua.Xiao@UTSouthwestern.edu}

  Jeffrey D. Allen \email{Jeffrey.Allen@UTSouthwestern.edu}
}
\keyword{datasets}

